+++ src/components/header/header_alert/header_alert.tsx
@@ -1,19 +1,24 @@
 import React, { FunctionComponent, HTMLAttributes, ReactNode } from 'react';
 import classNames from 'classnames';
 
-import { EuiButtonIcon } from '../../button';
 import { CommonProps } from '../../common';
 
 import { EuiFlexGroup, EuiFlexItem } from '../../flex';
-
-import { EuiI18n } from '../../i18n';
+import makeId from '../../form/form_row/make_id';
 
 export type EuiHeaderAlertProps = CommonProps &
   Omit<HTMLAttributes<HTMLDivElement>, 'title'> & {
+    /**
+     * Adds a link to the alert.
+     */
     action?: ReactNode;
     date: ReactNode;
     text?: ReactNode;
     title: ReactNode;
+    /**
+     * Accepts an `EuiBadge` that displays on the alert
+     */
+    badge?: ReactNode;
   };
 
 export const EuiHeaderAlert: FunctionComponent<EuiHeaderAlertProps> = ({
@@ -22,36 +27,27 @@ export const EuiHeaderAlert: FunctionComponent<EuiHeaderAlertProps> = ({
   date,
   text,
   title,
+  badge,
   ...rest
 }) => {
   const classes = classNames('euiHeaderAlert', className);
 
+  const ariaId = makeId();
+
   return (
-    <EuiI18n token="euiHeaderAlert.dismiss" default="Dismiss">
-      {(dismiss: string) => (
-        <div className={classes} {...rest}>
-          <EuiButtonIcon
-            aria-label={dismiss}
-            iconType="cross"
-            size="s"
-            className="euiHeaderAlert__dismiss"
-          />
+    <article aria-labelledby={`${ariaId}-title`} className={classes} {...rest}>
+      <EuiFlexGroup justifyContent="spaceBetween">
+        <EuiFlexItem>
+          <div className="euiHeaderAlert__date">{date}</div>
+        </EuiFlexItem>
+        {badge && <EuiFlexItem grow={false}>{badge}</EuiFlexItem>}
+      </EuiFlexGroup>
 
-          <div className="euiHeaderAlert__title">{title}</div>
-
-          <div className="euiHeaderAlert__text">{text}</div>
-
-          <EuiFlexGroup justifyContent="spaceBetween">
-            <EuiFlexItem grow={false}>
-              <div className="euiHeaderAlert__action euiLink">{action}</div>
-            </EuiFlexItem>
-
-            <EuiFlexItem grow={false}>
-              <div className="euiHeaderAlert__date">{date}</div>
-            </EuiFlexItem>
-          </EuiFlexGroup>
-        </div>
-      )}
-    </EuiI18n>
+      <h3 id={`${ariaId}-title`} className="euiHeaderAlert__title">
+        {title}
+      </h3>
+      <div className="euiHeaderAlert__text">{text}</div>
+      {action && <div className="euiHeaderAlert__action euiLink">{action}</div>}
+    </article>
   );
 };